grpc_server/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
httpx==0.27.0
sqlalchemy==2.0.25
googleapis-common-protos==1.62.0
protobuf>=4.25.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
psycopg2-binary==2.9.9